+++ src/main/java/_4352_4421_4480/springbootproject/student/StudentRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface StudentRepository extends JpaRepository<Student, Integer>
+public interface StudentRepository extends JpaRepository<Student, Long>
 {
-    Optional<Student> findById(int id);
+    Optional<Student> findById(Long id);
 }
